Add code tab DOM. Edit EAT and SLEEP split repetead code to module card-grid.
