+++ src/index.js
@@ -2,6 +2,7 @@ import {PAGELOAD, mobileTitleContainer, TABCONTENTS} from './modules/template'
 import {LOADHOME} from './modules/home'
 import {LOADEAT} from './modules/eat'
 import {LOADSLEEP} from './modules/sleep'
+import {LOADCODE} from './modules/code'
 import './main.scss';
 
-LOADSLEEP();
+LOADCODE();
